refactor: finalize electric tick boundary settlement

+++ moorestech_server/Assets/Scripts/Tests/CombinedTest/Core/ElectricToGearGeneratorTest.cs
@@ -38,14 +38,14 @@ namespace Tests.CombinedTest.Core
 
             // 初期はバッテリーが空なので満容量分を要求し、出力はしない
             // The empty battery initially demands its full capacity and produces no output
-            Assert.AreEqual((float)mode0.RequiredPower, c.RequestEnergy.AsPrimitive(), 0.001f);
+            Assert.AreEqual(mode0.RequiredPower, c.RequestEnergy.AsPrimitive(), 0.001f);
             Assert.AreEqual(0f, c.GenerateRpm.AsPrimitive(), 0.001f);
 
             // 供給率1の電力tickで満充電になり、定格RPM・定格トルクを出力する
             // A rate-1 electric tick fully charges the battery, outputting the rated RPM and torque
             c.OnElectricTickPostProcess(FullRateStats(mode0.RequiredPower));
-            Assert.AreEqual((float)mode0.Rpm, c.GenerateRpm.AsPrimitive(), 0.001f);
-            Assert.AreEqual((float)mode0.Torque, c.GenerateTorque.AsPrimitive(), 0.01f);
+            Assert.AreEqual(mode0.Rpm, c.GenerateRpm.AsPrimitive(), 0.001f);
+            Assert.AreEqual(mode0.Torque, c.GenerateTorque.AsPrimitive(), 0.01f);
 
             // 出力tickで1tick分のバッテリーを全消費し残量0・出力停止
             // The output tick consumes the whole one-tick battery, stopping the output at zero remainder
@@ -57,12 +57,12 @@ namespace Tests.CombinedTest.Core
             // At rate 0.5 the battery charges halfway and requests only the remaining half next tick
             c.OnElectricTickPostProcess(HalfRateStats(mode0.RequiredPower));
             Assert.AreEqual(0f, c.GenerateTorque.AsPrimitive(), 0.01f);
-            Assert.AreEqual((float)mode0.RequiredPower * 0.5f, c.RequestEnergy.AsPrimitive(), 0.001f);
+            Assert.AreEqual(mode0.RequiredPower * 0.5f, c.RequestEnergy.AsPrimitive(), 0.001f);
 
             // 残量要求を全量供給すると満充電になり、要求は0になる
             // Fully supplying the remaining demand completes the charge and reduces demand to zero
             c.OnElectricTickPostProcess(FullRateStats(c.RequestEnergy.AsPrimitive()));
-            Assert.AreEqual((float)mode0.Torque, c.GenerateTorque.AsPrimitive(), 0.01f);
+            Assert.AreEqual(mode0.Torque, c.GenerateTorque.AsPrimitive(), 0.01f);
             Assert.AreEqual(0f, c.RequestEnergy.AsPrimitive(), 0.001f);
             c.ConsumeGeneratorTick(1f);
 
@@ -70,14 +70,14 @@ namespace Tests.CombinedTest.Core
             // Switching modes while partially charged updates the demand and keeps the output off until fully charged
             c.OnElectricTickPostProcess(HalfRateStats(mode0.RequiredPower));
             c.SetSelectedMode(1);
-            Assert.AreEqual((float)mode1.RequiredPower - (float)mode0.RequiredPower * 0.5f, c.RequestEnergy.AsPrimitive(), 0.001f);
+            Assert.AreEqual(mode1.RequiredPower - mode0.RequiredPower * 0.5f, c.RequestEnergy.AsPrimitive(), 0.001f);
             Assert.AreEqual(0f, c.GenerateTorque.AsPrimitive(), 0.01f);
 
             // 供給率1の電力tickで満充電になり、新モードの定格を出力（トルクドループなし）
             // A rate-1 electric tick fills the battery and outputs the new mode rated values (no torque droop)
             c.OnElectricTickPostProcess(FullRateStats(c.RequestEnergy.AsPrimitive()));
-            Assert.AreEqual((float)mode1.Torque, c.GenerateTorque.AsPrimitive(), 0.01f);
-            Assert.AreEqual((float)mode1.Rpm, c.GenerateRpm.AsPrimitive(), 0.001f);
+            Assert.AreEqual(mode1.Torque, c.GenerateTorque.AsPrimitive(), 0.01f);
+            Assert.AreEqual(mode1.Rpm, c.GenerateRpm.AsPrimitive(), 0.001f);
 
             Assert.IsFalse(c.SetSelectedMode(99));
             Assert.AreEqual(1, c.SelectedIndex);
@@ -156,7 +156,7 @@ namespace Tests.CombinedTest.Core
             // index 3 is the torque-0 mode; supply it to full fulfillment.
             var mode3 = param.OutputModes[3];
             c.SetSelectedMode(3);
-            c.OnElectricTickPostProcess(new ElectricNetworkStatistics((float)mode3.RequiredPower, (float)mode3.RequiredPower, 1f, 1));
+            c.OnElectricTickPostProcess(new ElectricNetworkStatistics(mode3.RequiredPower, mode3.RequiredPower, 1f, 1));
 
             // 満充足でもトルク0なら GenerateRpm は0（実効トルクゲート）。網の最速起点を奪わない。
             // Even at full fulfillment, a torque-0 mode yields GenerateRpm 0 (effective-torque gate); never dominates the network.
